Healthd: Correct the code logic of props.batteryPresent
In healthd, mAlwaysPluggedDevice indicates the type of power supply of
a platform. Specifically, the value of mAlwaysPluggedDevice means:
(1) "false", a real battery is used;
(2) "true", battery is not present and an external power-supply used.

However, in following code section, it set "props.batteryPresent = true"
when mAlwaysPluggedDevice is "true" which is a paradox design. As a
result, healthd and Android will still try a battery status checking
and update even if an external power supply is used indeed. And battery
status will be always in "charging status" which blocks some CTS tests.

This patch corrects the code logic error and set right properties of
battery and charger for the always plugged device.

Bug: 70648443
Test: Pass CTS Network Connectivity Screen Off Test with AlwaysPluggedDevice
Change-Id: I28d73734c6466901c4c0f0d6a71fc5ffaa5f2cf3
Signed-off-by: Ren Yanx <yanx.ren@intel.com>
Signed-off-by: Richard JK <rjk@codeaurora.org>