libsuspend: Add property support for timeout of autosuspend.
THe implementation to increase the delay time where system attempt to
suspend whenever suspend fails. The delay keeps on increasing for every
consecutive suspend failure. The aim of this feature is to prevent
device from entering suspend if they are active wakeup events happens on
the device in the absence of a wakelock.

The max delay timeout it today hard coded to 60 seconds and the value is
not optimized across different platforms. This value should ideally be
equal to the value where the power cost of attempting suspend matches to
the overhead of staying in CPU idle. This breakeven can vary based on
the chipset and the software that the oem add to the device. So we would
like to make the max timeout value to be configurable based on a
property.

Change-Id: I0ba1cf782f2915c51117bf592cfe43ca69f3b223