adbd: Fix USB enumeration on adbd restart
On latest targets, boot usb composition is stored in /etc
partition. But the ADB client checks for it in the /data
partition where composition symlink was stored earlier and hence
fails to write udcname to UDC, resulting in USB enumeration
failure on adbd restart.
Fix this by not checking for boot usb composition because restart
of adbd will not happen before start_usb script is run which
would definitely populate boot usb composition info, be it in
/data or /etc partition.

Change-Id: If4826e0d65b28a8d47d37389fbf15a2ec7b9ae7a
Signed-off-by: Ajay Agarwal <ajaya@codeaurora.org>