Revert "system: core: usb: Add delay between stop/start adbd"
Commit 173e5b9d0436 added a delay between restart of ADBD and
binding of functions to resolve the issue of race between
composition bind from script rules and restart of ADBD which
unbinds the compositions. But ideally kill and start of ADBD
should unbind and bind the functions and re-enumeration should
happen. Such delay is not needed once that is fixed in the ADB
client. Hence, revert commit 173e5b9d0436.

Change-Id: Ifa66cf55b65592eaca322c6389546a4d5f19a6e3