adbd: Add changes for binding to UDC from adbd
Drivers support to bind udc with configfs when adbd is
started/restarted. This is required as adbd uses function fs
driver which fails the fs bind when adbd is not ready.

In the current code ADB client checks boot composition in the
/data partition where composition symlink was stored earlier,
and sys.usb.configfs property and fails to write udcname to UDC,
resulting in USB enumeration failure on adbd restart.

Fix this by not checking for boot usb composition in /data instead
added bind_in_progress in composition scripts to check whether
bind from userspace composition script happen or not. If not then
explicitly bind UDC from the adbd.

Change-Id: Ia714365a5293e634afa5e181bf990609286f5a83