usb: composition: Correct subCompatible ID passing for MBIM function.
Composite gadget driver tells windows about the number of config on
which MBIM is bound it by returning the num as a string in the OS descriptor's
subCompatibleID field. Currently we are passing incorrect subCompatible
string from user space for MBIM function in multi-config compositions.
This is causing issues with multi-config support where windows10/8 host's does
not select the correct configuration.

Hence fix this issue by correcting the string passing for OS
descriptor's subCompatibleID.

Change-Id: I542c3f8bf8990a1d2ecdfe58ae8047596cf77c82