Revert "usb: Set primary controller to peripheral for SM8150 automotive"
Revert commit 1a01acaddad0. There is support in dwc3-msm driver
that if extcon is not present and dr_mode is OTG, then the USB
controller is set to peripheral mode. Since extcon is not yet
present for Auto targets, it is not required to set default
peripheral mode from userspace.

Change-Id: Id04e05c64a4b92bc5c7f07070690d4eb2a30744c