usb: composition: Add configfs based rules for 90B1 composition
Add USB configfs based rules with 90B1 USB composition which allows
to use this composition on newer platform where configfs is used
to create USB composite device. Current code only allows usage of
this composition with HSIC interface only which prevents running
this rules when USB is selected as interface. Fix this so this
composition can be used with USB interface.

Change-Id: I61c7beb234c1ab9e0775122ebb9abfb61e9bd609