system: core: usb: Remove remote wakeup capability on bmAttributes
USB functions in compositions 904A and 9060 does not support remote
wakeup. But the bmAttributes for these composition have the remote
wakeup capability set leading to failure of Remote Wakeup tests of
Chapter 9 compliance in HighSpeed.
Hence remove setting the remote wakeup capability (bit-5) of the
bmAttributes to fix the failures.

Change-Id: I0ed30a881783e09d2c02a857c9e1b62562611a56