system: core: usb: Add print statement at the end of start_usb
Currently the last executable line of code runs the composition
script which is populated in the pid variable. Let's assume,
some other script tries to populate illegal PID to the pid variable.
Now start_usb script will try to find the composition script of the
PID and if not found, it will return an error exit code to the caller
of the script. This leads to failure in starting usb.service. To avoid
this add a print statement so that script never returns an error exit
code.

Change-Id: I5486eb63414eae048853cd322561343566814b82
Signed-off-by: Pratham Pratap <prathampratap@codeaurora.org>