system: core: usb: Update diag dload with PID and serial
With support now for diag-router using USB FFS driver, the kernel
driver is no longer going to automatically update the DLOAD IMEM
with the required PID/Serial information that is used when the device
transitions into crash download mode so that a port re-enumeration
is avoided. In order to pass the same PID/Serial information, the
same kernel driver now exposes configfs attributes that userspace
can write. Hence the serial number can be written once during start_usb,
and likewise the PID can be updated whenever a composition
is selected that contains Diag as the first interface.

Change-Id: I96c86ae4b66ea3bedd7a0951ae78fcf56fa6aa36
Signed-off-by: Elson Roy Serrao <eserrao@codeaurora.org>