system: core: usb: Write MAC Address to usb_qti partition
For early USB enumeration the static MAC adress is not available
in usb_qti partition and due to that ECM data is taking SW path
instead of IPA HW path. currently the magic number, PID and
product string is present in the usb_qti partition and that ABL
will read it and pass it to kernel via kernel cmdline on
subsequent reboot, so add the same support for the MAC_address.

To update the MAC adress in usb_qti partition user should write
the MAC adress in below path,
echo “MAC add” > /sys/module/usb_f_gsi/parameters/gsi_host_addr

Change-Id: I62c025e95936dcc8f709b2f997a6d23cd5ea46e1
Signed-off-by: Aniket Randive <arandive@codeaurora.org>