adbd: Add AVB2.0 support for dm-verity
When AVB2.0 is used, ADB is used for disabling/enabling
dm-verity in user builds.

Prior to this change, only AVB1.0 was supported, where verity
information is stored in the kernel cmdline in boot image header.
With AVB2.0, cmdline with verity information is passed in vbmeta.

This change adds a new flag "ADB_AVB20", to enable this behavior.

Signed-off-by: Amit Blay <ablay@codeaurora.org>
Change-Id: I04a033bba0fc3170c021f9c5b7cc82fc8fe70909